- Rename controller to reflect what it actually does :)
